New Features
A lot of new features now are supported.
I've also reduced variable name lengths to reduce the KB count. (Only in manager and projection files)

It's too much to talk about so I won't explain absolutely everything that was added. All the features will be explained in the documentation coming later.