Added Action-Event Processing Back
Added back in action-event processing. Still potentially broken, but should function fine. Replaced the gx, gy, gz with px, py, pz, which is the position relative to the center of the element in question. More applicable in most cases than gx, gy, and gz and still easy to convert to global coordinates, in theory, probably.